app/requirements.txt
torch
torchvision
matplotlib
streamlit
pillow
numpy
psycopg2-binary
python-dotenv
